fully refactored code to use rsc patterns
