+++ apps/admin-portal/src/app/_components/order/CommentsContainer.tsx
@@ -1,0 +1,27 @@
+"use client";
+
+import { api } from "../../../trpc/react";
+import CommentForm from "./CommentForm";
+
+interface CommentsContainerProps {
+  orderId: number;
+  locationId: number;
+}
+
+export default function CommentsContainer({
+  orderId,
+  locationId,
+}: CommentsContainerProps) {
+  const { data: locationEmployees } =
+    api.orderComments.getLocationEmployees.useQuery({
+      locationId: locationId,
+    });
+
+  return (
+    <CommentForm
+      orderId={orderId}
+      locationId={locationId}
+      locationEmployees={locationEmployees || []}
+    />
+  );
+}
