+++ packages/admin-api/src/root.ts
@@ -1,6 +1,5 @@
 import { analyticsRouter } from "./router/analytics";
 import { authRouter } from "./router/auth";
-import { customerUserRouter } from "./router/customer-user";
 import { notification } from "./router/notification";
 import { orderComments } from "./router/order-comments";
 import { ordersRouter } from "./router/orders";
@@ -13,7 +12,6 @@ export const appRouter = createTRPCRouter({
   user: userRouter,
   orderComments: orderComments,
   notification: notification,
-  customerUser: customerUserRouter,
   analytics: analyticsRouter,
 });
